+++ src/Entity.cpp
@@ -12,7 +12,7 @@ using namespace std;
 
 Logger Entity::logger = Logger::getLogger("Entity");
 
-Entity::Entity(Sprite* sprite, uint16_t width, uint16_t height) {
+Entity::Entity(Sprite* sprite, uint32_t width, uint32_t height) {
 	this->sprite = sprite;
 	// NOTE: are values of `SDL_Rect` implicitly set to 0 by default?
 	this->rect.x = 0;
@@ -35,24 +35,24 @@ Entity::Entity(Sprite* sprite) {
 	}
 }
 
-Entity::Entity(std::string sprite_id, uint16_t width, uint16_t height):
+Entity::Entity(std::string sprite_id, uint32_t width, uint32_t height):
 		Entity(SpriteStore::get(sprite_id), width, height) {}
 
 Entity::Entity(string sprite_id): Entity(SpriteStore::get(sprite_id)) {}
 
 
-bool Entity::collides(uint16_t x, uint16_t y, uint16_t len, bool horizontal) {
+bool Entity::collides(uint32_t x, uint32_t y, uint32_t len, bool horizontal) {
 	// this entity's right-most & bottom-most pixel collision
-	uint16_t right = this->rect.x + this->rect.w;
-	uint16_t bottom = this->rect.y + this->rect.h;
+	uint32_t right = this->rect.x + this->rect.w;
+	uint32_t bottom = this->rect.y + this->rect.h;
 
 	if (horizontal) {
-		uint16_t x_max = x + len;
+		uint32_t x_max = x + len;
 		if ((bottom >= y && this->rect.y <= y) && (right >= x && this->rect.x <= x_max)) {
 			return true;
 		}
 	} else {
-		uint16_t y_max = y + len;
+		uint32_t y_max = y + len;
 		if ((right >= x && this->rect.x <= x) && (bottom >= y && this->rect.y <= y_max)) {
 			return true;
 		}
@@ -63,13 +63,13 @@ bool Entity::collides(uint16_t x, uint16_t y, uint16_t len, bool horizontal) {
 bool Entity::collides(Entity* other) {
 	SDL_Rect o_rect = other->getRect();
 	// other entity's right-most & bottom-most pixel collision
-	uint16_t o_right = o_rect.x + o_rect.w;
-	//uint16_t o_top = o_rect.y;
-	uint16_t o_bottom = o_rect.y + o_rect.h;
+	uint32_t o_right = o_rect.x + o_rect.w;
+	//uint32_t o_top = o_rect.y;
+	uint32_t o_bottom = o_rect.y + o_rect.h;
 
 	// this entity's right-most & bottom-most pixel collision
-	uint16_t t_right = this->rect.x + this->rect.w;
-	uint16_t t_bottom = this->rect.y + this->rect.h;
+	uint32_t t_right = this->rect.x + this->rect.w;
+	uint32_t t_bottom = this->rect.y + this->rect.h;
 
 	// check for clipping at any point
 	return ((this->rect.x >= o_rect.x && this->rect.x <= o_right)
